requirements/common.txt
bs4
sphinx
click
ffmpy
scipy
numpy
matplotlib
lxml
google-api-python-client
oauth2client
sklearn
pymongo
python-dotenv
boto3
editdistance
tqdm
minio
profilehooks
